feat: add module selector page
