+++ .env.example
@@ -1,3 +1,5 @@
+VITE_APP_NAME=
+
 VITE_API_KEY=your_api_key
 VITE_AUTH_DOMAIN=your_auth_domain
 VITE_PROJECT_ID=your_project_id
