--- src/layouts/BaseLayout/index.ts
@@ -1,1 +1,0 @@
-export { default as BaseLayout } from "./Layout";
