+++ src/lib/consts.ts
@@ -1,5 +1,6 @@
 import packageJson from "../../package.json";
 
+export const APP_NAME = import.meta.env.VITE_APP_NAME;
 export const VERSION = import.meta.env.VITE_APP_VERSION || packageJson.version;
 
 export const LOCAL_STORAGE_KEYS = {
@@ -17,4 +18,5 @@ export const ROUTES = {
       IMPERSONATE: "/inventory/admin/impersonate",
     },
   },
+  SALES: "/sales",
 };
