+++ src/main.tsx
@@ -4,6 +4,7 @@ import { RouterProvider } from "react-router-dom";
 
 import { AuthProvider } from "./context";
 import { router } from "./lib/routes.tsx";
+import "@fontsource/roboto";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   // <React.StrictMode>
